A Custom ggplot2 theme for the TCPS

Custom ggplot2 theme for the TCPS Package heavy credits for influencing the theme function go to @hrbrmstr (Bob Rudis)

theme_tcps(base_family = "mono", base_size = 11,
  strip_text_family = base_family, strip_text_size = 12,
  plot_title_family = "mono", plot_title_size = 16,
  plot_title_margin = 10, subtitle_family = "mono",
  subtitle_size = 12, subtitle_margin = 15, caption_family = "serif",
  caption_size = 9, caption_margin = 10, axis_title_family = "serif",
  axis_title_size = 9, axis_title_just = "mm", grid = TRUE,
  axis = FALSE, ticks = FALSE)

Arguments

base_family
base font family

base_size
base font size

strip_text_family
facet label font family

strip_text_size
facet label text size

plot_title_family
plot tilte family

plot_title_size
plot title font size

plot_title_margin
plot title margin

subtitle_family
plot subtitle family

subtitle_size
plot subtitle size

subtitle_margin
plot subtitle margin

caption_family
plot caption family

caption_size
plot caption size

caption_margin
plot caption margin

axis_title_family
axis title font family

axis_title_size
axis title font size

axis_title_just
axis title font justification blmcrt

grid
panel grid (TRUE, FALSE, or a combination of X, x, Y, y)

axis
axis TRUE, FALSE, xy

ticks
ticks

Details

Proper display of the theme requires installing Oswald & Oxygen fonts unless you change the font parameters

https://www.google.com/fonts